Modified openFileContainingVar to work on all platforms.
The Clojure openFileContainingVar script used to fail on Windows,
because it did not handle the path in a multiplatform safe way and it
was relying on the unzip external utility that does not exist on
Windows.

I changed the code to generate valid paths no matter what platform we're
running on by relying on Java's path manipulation constructs instead of
string concatenation.

I also modified the unziping of the Jar to be performed all in Clojure
instead of relying on an external tool.

These changes were made to fix the "proto-repl:open-file-containing-var"
command on Windows.

Issues Related:
- https://github.com/jasongilman/proto-repl/issues/172
- https://github.com/jasongilman/proto-repl/issues/217