Add Gradle support
The gradle-clojure plugin [1] is adding Clojure support to Gradle
and now (as of 0.3.0-rc.2) supports starting an nREPL server. This
adds support for Gradle in addition to the existing Boot and
Leiningen support.

The Gradle support does not specify a gradlePath config setting, like
Lein and Boot do, since idiomatic Gradle usage uses a wrapper shell
launcher included within the project. As such, we search for that
gradlew file instead of a build.gradle to determine if it supports
Gradle.

Previously, there was a boolean preferLein option to choose between
Lein and Boot, if both were present. With a third tool, this gets a
little more complicated. This is now supported via an array of
enumerated values to indicate order of preference.

There may be a simpler approach to get at the gist of the feature.

[1] https://github.com/gradle-clojure/gradle-clojure