Added support for catching exceptions thrown by a Boot REPL
Previously, any code executed in a Boot REPL (remote or local) that threw an exception would cause a spinner to appear in the editor and never disappear. The problem is, Boot returns exceptions in a slightly different way, causing proto-repl to ignore the returned message and never trigger the result handler which would cleanup the spinner.

This now checks for Boot exception messages and calls the result handler fixing the issue.

Fixes issues #209 and #215 for local and remote REPLs.